unify prod/dev add-on build
