+++ package.json
@@ -54,10 +54,10 @@
     "broccoli-yuidoc": "^2.1.0",
     "ember-ajax": "0.7.1",
     "ember-cli": "2.3.0",
-    "ember-cli-app-version": "0.5.0",
+    "ember-cli-app-version": "^1.0.0",
     "ember-cli-blueprint-test-helpers": "^0.9.0",
     "ember-cli-dependency-checker": "^1.2.0",
-    "ember-cli-htmlbars": "0.7.9",
+    "ember-cli-htmlbars": "^1.0.3",
     "ember-cli-htmlbars-inline-precompile": "^0.2.0",
     "ember-cli-inject-live-reload": "^1.3.1",
     "ember-cli-internal-test-helpers": "^0.8.1",
